requirements.txt
amqp==2.2.2
astroid==1.6.0
billiard==3.5.0.3
celery==4.1.0
Django==2.0
djangorestframework==3.7.3
isort==4.2.15
kombu==4.1.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.8.1
pymongo==2.8
pytz==2017.3
redis==2.10.6
six==1.11.0
vine==1.1.4
virtualenv==15.1.0
wrapt==1.10.11
